Adding tests and make rng seed settable for reproduciblity
Okay. Now we would have an easier time making changes if we had tests in place, so please right some tests. We'd also find ourselves having an easier time if we had a centralized source of randomness - so make it possible to set the seed for the deck and use it consistently, not always resetting the seed.