Adds play/discard and targeting 300 points as gameplay concepts
Okay. We've demonstrated the ability to score hands. Let's add a bit more gameplay. We also have discards. I should be able to select cards and either discard them or play them (as a hand).

While we're enhancing the gameplay loop, let's change from an infinite loop of scoring hands, and move to a single challenge of scoring 300 points. I should have 4 hands to play and 3 discards to get to 300 points