Adds the economic system and the first joker!
Now we need to introduce 3 related concepts: money, rewards, the shop, and our first Joker.

A player has money. They earn it primarily through completing Blinds (though we'll add support for other ways of earning money) and spend it in the shop. We need to keep track of money throughout the game.

After completing a blind, the player earns rewards. They earn $4, $5, $6 for completing the small, big, and boss blinds respectively. _Plus_, they earn $1 per unused hand and $1 per unused discard.

The shop appears between each Blind.There, the player gets the opportunity to spend money. For now, the only thing the player can spend money on is

The first Joker. This is The Golden Joker, which grants $4 at the end of each Blind. Go for it!