Makes scoring tweakable via csv
I'd like to be able to make gameplay balance changes without recompiling. Move the ante score requirements to a CSV (each row represents an ante, and there's 3 columns - small, big, boss). Move the hand scores to a CSV as well (each row represents a hand. headers are base and mult)