Big failure here. Zed kept getting caught in infinite loops.
It wants to test by piping input into a game it starts, but at this point the game is too complex to handle that way.
I might have to get it to stop running the game and start relying on tests.

A requirement I failed to mention: the shop has two purchases available that are randomly selected from the pool of all purchaseable items. If you don't like those items (or you do like them and purchase them both) you can (r)eroll. Rerolling starts at $5 and goes up by $1 for each time it's performed. The price resets when you start a new blind.