Still no good. Did a bunch of work on the TUI but it's not static at all.
I'd like the TUI to feature lots of staticly layed-out data. The Ante, the Blind, the money available and the jokers in play should never move. The cards in hand should always be in the same spot, as should the command prompt. Rerender over old data, rather than adding information to another part of the screen.