Adds a game io interface to separate concerns around game vs rendering
Okay. The TUI is working well now. But we have a problem both with the TUI and the logger mode of playing the game. The game is not factored separately from the input/output logic.  In both places we end up with deeply intermingled logic. I'd like to refactor the game such that it handle IO through an interface that can support either the logger mode or the TUI mode.